Pull request #3548: HTTP/3 inspector implementation support
Merge in SNORT/snort3 from ~ABHPAL/snort3:h3 to master

Squashed commit of the following:

commit 18d340b34fb619533c4a8d1722cd57f823d817ba
Author: Abhijit Pal(abhpal) <abhpal@cisco.com>
Date:   Thu Aug 25 16:21:19 2022 +0530

    parser: add implicit http3 to http ips options otn

commit b38f067a20e4503d29916be966919fafee71d3c7
Author: Abhijit Pal(abhpal) <abhpal@cisco.com>
Date:   Thu Aug 25 16:20:14 2022 +0530

    stream: export support for creating udp session

commit 6f3f7109f8f3c8b0c3299a2aec7c58508a000840
Author: Abhijit Pal(abhpal) <abhpal@cisco.com>
Date:   Thu Aug 25 16:18:56 2022 +0530

    detection: add http3 to http ips buffers

commit 254ccfed242e89b5780407691c5b9fff69684be4
Author: Abhijit Pal(abhpal) <abhpal@cisco.com>
Date:   Thu Jul 14 15:54:39 2022 +0530

    flow: abstract class added to work on stream based connections

commit d2b82a8feccd6ac3c37aa202ec58c505714f7546
Author: Abhijit Pal(abhpal) <abhpal@cisco.com>
Date:   Thu Jul 14 15:53:56 2022 +0530

    pub_sub: handle httpx(2,3) traffic

commit 9bf0c34a118bd4f3dba8052ee141be1a86eea237
Author: Abhijit Pal(abhpal) <abhpal@cisco.com>
Date:   Thu Jul 14 15:52:55 2022 +0530

    payload_injector: accomodate httpx(2,3) stream id values

commit 32e13e3f1f534f5632264e3e0d1d9f1f921c74b8
Author: Abhijit Pal(abhpal) <abhpal@cisco.com>
Date:   Thu Jul 14 15:52:06 2022 +0530

    rna: handle httpx(2,3) traffic

commit c3ad5f625c98a337d2bf5b51742075d1d5b07c23
Author: Abhijit Pal(abhpal) <abhpal@cisco.com>
Date:   Thu Jul 14 15:51:23 2022 +0530

    appid: handle http event for httpx(2,3) traffic

commit b7e9927040da7d01ebb3dbed0b256340a5bf4f94
Author: Abhijit Pal(abhpal) <abhpal@cisco.com>
Date:   Thu Jul 14 15:50:20 2022 +0530

    http2_inspect: updated with abstracted httpx(2,3) flags

commit d27580f9f0666ec765c90347a34ccad619effcb0
Author: Abhijit Pal(abhpal) <abhpal@cisco.com>
Date:   Thu Jul 14 15:48:46 2022 +0530

    http_inspect: abstract inspection of httpx(2,3)

+++ src/payload_injector/payload_injector.h
@@ -48,7 +48,7 @@ struct InjectionControl
 {
     const uint8_t* http_page = nullptr;
     uint32_t http_page_len = 0;
-    uint32_t stream_id = 0;
+    int64_t stream_id = 0;
 };
 
 class SO_PUBLIC PayloadInjector
